Move config related tests
