+++ python/jupyterlab-collaborative-chat/ui-tests/tests/unread.spec.ts-snapshots/navigation-top-linux.png
